factor common elements in AnalyzeProjectSitesMapCard and ResilienceMapCard into MapCard
